+++ CHANGELOG.md
@@ -27,6 +27,7 @@ to know to migrate from 0.22.x release.
 - The `generation_time` of the report is now computed by the reporting backends
 - JSON/XML: dates are now stored in plain ISO8601 format in UTC
 - HTML: dates are now localized in the browser's timezone
+- HTML: statistics are computed in Javascript
 
 ## Under the hood