+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.22.10 (2019-05-19)
+
+- **lcc run**: display skipped tests as '--' instead of 'KO'
+
 # 0.22.9 (2019-04-28)
 
 - **Report**: provide more information about why a test has been skipped